remove broken `binaryFloat` and replace with `floating`
`floating` is hopefully correct.  I've tested `encode`/`decode` a
bunch including with QuickCheck, and tested it on a lot of examples,
but I can't be absolutely sure.  It's kind of cool that we can find
the precise floating point numbers closest to the boundary, but
otherwise this seems pretty useless. =)